New dto files created
